requirements.txt
azure-functions
black
pandas
sklearn
tensorflow